+++ main/tags/tagger.js
@@ -468,10 +468,11 @@ function Tagger({
 		this.debouncedStep({ step: this.iStep, bDebug, bProfile });
 	};
 
-	const orderKeys = ['rgScan', 'tpScan', 'biometric', 'masstagger', 'dynamicRange', 'drMeter', 'chromaPrint', 'ffmpegLRA', 'folksonomy', 'essentiaFastKey', ['essentiaKey', 'essentiaBPM', 'essentiaDanceness', 'essentiaLRA']]; // Must follow order at this.stepTag()
+	const orderKeys = ['rgScan', 'tpScan', 'biometric', 'masstagger', 'dynamicRange', 'drMeter', 'chromaPrint', 'ffmpegLRA', 'folksonomy', 'essentiaFastKey', ['essentiaKey', 'essentiaBPM', 'essentiaDanceness', 'essentiaLRA'], 'audioMd5', 'rgScan', 'tpScan', 'bpmAnaly']; // Must follow order at this.stepTag()
 	if (orderKeys.flat(Infinity).some((k) => !Object.hasOwn(this.toolsByKey, k))) { throw new Error('Key not associated to any tool'); }
 	this.stepTag = ({ step, bDebug = false, bProfile = false }) => {
 		const runMenu = (menuArr, handleList, title) => {
+			if (!handleList) { return false; }
 			bSuccess = menuArr.some((name) => fb.RunContextCommandWithMetadb(name, handleList, 8));
 			if (!bSuccess) { fb.ShowPopupMessage('Contextual menu entries not found:\n\n  - ' + menuArr.join('\n  - ') + '\n\nCheck they match the contextual menus associated to the component and don\'t have any typo. Otherwise report to the component\'s dev.', title); }
 			return bSuccess;
@@ -616,7 +617,7 @@ function Tagger({
 	this.debouncedStep = debounce(this.stepTag, this.timers.debounce); // Only continues next step when last tag update was done > X ms ago
 
 	this.checkHandleList = ({ bDebug = false, bProfile = false } = {}) => {
-		if (!this.isRunning) { this.stopStepTag(); return; }
+		if (!this.isRunning()) { this.stopStepTag(); return; }
 		const i = this.iStep - 1;
 		if (i >= 0) {
 			const key = orderKeys[i];
@@ -628,8 +629,7 @@ function Tagger({
 				const mergeTags = new Set(['chromaPrint', 'folksonomy']);
 				const handleList = this.notAllowedTools.has(key) ? this.selItemsByCheck[checkKey].missing : this.selItems;
 				if (this.toolsByKey[key] && handleList.Count) {
-					const idx = this.tools.findIndex((tool) => { return tool.key === key; });
-					const tag = this.tools[idx].tag;
+					const tag = this.tagsByKey[key];
 					const itemTags = getHandleListTags(handleList, tag, { bMerged: true })
 						.map((tagArr) => mergeTags.has(key) ? tagArr.join(', ') : tagArr)
 						.flat(Infinity).filter(Boolean);